+++ packages/database/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Unreleased
 
+## 0.12.6
+
+### Patch Changes
+
+- Updated dependencies [[`2d672cead`](https://github.com/firebase/firebase-js-sdk/commit/2d672cead167187cb714cd89b638c0884ba58f03)]:
+  - @firebase/util@1.5.0
+  - @firebase/component@0.5.11
+
 ## 0.12.5
 
 ### Patch Changes